split vaccine and scheduleVaccine test files
